Refs #4, move utils.py up, cleaning the structure, some gitignore additions
